# internal/applications/rest/data/openapi.yaml
openapi: 3.1.0
info:
  title: Ovoo API
  description: >-
    Ovoo - privacy email gateway to manage aliases hiding real (protected)
    emails from senders.
  contact:
    email: burmuley@gmail.com
    name: Konstantin Vasilev
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: "1.2"
...
security:
  - OAuth2: []
  - BasicAuthentication: []
  - ApiToken: []
servers:
  - url: https://ovoolocal.burmuley.com:8808
    description: Local testing server
    variables: {}
  - url: https://ovoomail.burmuley.com:8808
    description: Public testting server
...
paths:
...
  /api/v1/users:
    get:
      description: >-
        Retrieve a list of users accessible for the currently authenticated
        user. In case if authenticated user is not of admin role -
        the only record in the list is the user information about itself (all filters are ignored)
      operationId: getUsers
      tags:
        - Users
      parameters:
        - in: query
          name: id
          description: user id filter
          schema:
            type: string
          required: false
        - in: query
          name: login
          description: user login filter
          schema:
            type: string
          required: false
        - in: query
          name: type
          description: user type filter
          schema:
            type: string
          required: false
      responses:
        "200":
          $ref: "#/components/responses/getUsersResponse"
        "401":
          $ref: "#/components/responses/HTTP401"
        "403":
          $ref: "#/components/responses/HTTP403"
        "404":
          $ref: "#/components/responses/HTTP404"
      security:
        - OAuth2: []
        - BasicAuthentication: []
      summary: Get all users
    post:
      description: >-
        Create new user. Only users if type \`admin\` are authorized to submit
        this request.
      operationId: createUser
      tags:
        - Users
      parameters: []
      requestBody:
        $ref: "#/components/requestBodies/createUserRequest"
      responses:
        "201":
          $ref: "#/components/responses/createUserResponse"
        "400":
          $ref: "#/components/responses/HTTP400"
        "401":
          $ref: "#/components/responses/HTTP401"
        "403":
          $ref: "#/components/responses/HTTP403"
      summary: Create user
      security:
        - OAuth2: []
        - BasicAuthentication: []
  /api/v1/users/{id}:
    get:
      description: >-
        Retrieve deteiled information of a particular user by its ID. The ID is
        defined as a path parameter and should be of ULID format.
      operationId: getUserDetails
      tags:
        - Users
      parameters: []
      responses:
        "200":
          $ref: "#/components/responses/getUserDetailsResponse"
        "401":
          $ref: "#/components/responses/HTTP401"
        "403":
          $ref: "#/components/responses/HTTP403"
        "404":
          $ref: "#/components/responses/HTTP404"
      security:
        - OAuth2: []
        - BasicAuthentication: []
      summary: Get user details
    patch:
      description: >-
        Update user information such as \`first\_name\`, \`last\_name\` and
        \`type\`. No other fields can be updated. The ID is defined as a path
        parameter and should be of ULID format.
      operationId: updateUser
      tags:
        - Users
      parameters: []
      requestBody:
        $ref: "#/components/requestBodies/updateUserRequest"
      responses:
        "200":
          $ref: "#/components/responses/updateUserResponse"
        "400":
          $ref: "#/components/responses/HTTP400"
        "401":
          $ref: "#/components/responses/HTTP401"
        "403":
          $ref: "#/components/responses/HTTP403"
        "404":
          $ref: "#/components/responses/HTTP404"
      summary: Update user information
      security:
        - OAuth2: []
        - BasicAuthentication: []
    delete:
      description: >-
        Delete a particular user by its ID. The ID is defined as a path
        parameter and should be of ULID format.
      operationId: deleteUser
      tags:
        - Users
      parameters: []
      responses:
        "204":
          description: ""
          content: {}
          headers: {}
        "401":
          $ref: "#/components/responses/HTTP401"
        "403":
          $ref: "#/components/responses/HTTP403"
        "404":
          $ref: "#/components/responses/HTTP404"
      summary: Delete user
      security:
        - OAuth2: []
        - BasicAuthentication: []
    parameters:
      - in: path
        name: id
        description: "User ID"
        schema:
          type: string
        required: true
  /api/v1/users/profile:
    get:
      summary: Get current user profile
      description: Retrieves currently logged on user profile
      operationId: getUserProfile
      tags:
        - Users
      parameters: []
      responses:
        "200":
          $ref: "#/components/responses/getUserDetailsResponse"
        "401":
          $ref: "#/components/responses/HTTP401"
...
  /api/v1/praddrs:
    get:
      summary: Get protected addresses available for the authenticated user
      description: >-
        Retrieve a list of protected addresses belonging to the currently
        authenticated user. Optional fields \`id\` and \`email\` can be used to
        limit the search scope.
      operationId: getPrAddrs
      tags:
        - Protected Addresses
      parameters:
        - in: query
          name: id
          description: ""
          schema:
            type: string
        - in: query
          name: email
          description: ""
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/getPrAddrsResponse"
        "401":
          $ref: "#/components/responses/HTTP401"
        "403":
          $ref: "#/components/responses/HTTP403"
        "404":
          $ref: "#/components/responses/HTTP404"
      security:
        - OAuth2: []
        - BasicAuthentication: []
    post:
      summary: >-
        Create new protected address and make current authenticated user its
        owner
      description: >-
        Create new protected address. Request is of the type
        \`application/json\`.
      operationId: createPrAddr
      tags:
        - Protected Addresses
      parameters: []
      requestBody:
        $ref: "#/components/requestBodies/createProtectedAddressRequest"
      responses:
        "200":
          $ref: "#/components/responses/createPrAddrResponse"
        "400":
          $ref: "#/components/responses/HTTP400"
        "401":
          $ref: "#/components/responses/HTTP401"
        "403":
          $ref: "#/components/responses/HTTP403"
      security:
        - OAuth2: []
        - BasicAuthentication: []
  /api/v1/praddrs/{id}:
    get:
      summary: Get details on a particular protected address
      description: >-
        Retrieve deteiled information of a particular protected address by its
        ID. The ID is defined as a path parameter and should of ULID format.
      operationId: getPrAddrDetails
      tags:
        - Protected Addresses
      parameters: []
      responses:
        "200":
          $ref: "#/components/responses/getPrAddrDetailsResponse"
        "401":
          $ref: "#/components/responses/HTTP401"
        "403":
          $ref: "#/components/responses/HTTP403"
        "404":
          $ref: "#/components/responses/HTTP404"
      security:
        - OAuth2: []
        - BasicAuthentication: []
    patch:
      summary: Update defined protected address by its ID
      description: >-
        Update protected address information such as \`metadata\`. No other
        fields can be updated. The ID is defined as a path parameter and should
        be of ULID format.
      operationId: updatePrAddr
      tags:
        - Protected Addresses
      parameters: []
      requestBody:
        $ref: "#/components/requestBodies/updateProtectedAddressRequest"
      responses:
        "200":
          $ref: "#/components/responses/updatePrAddrResponse"
        "400":
          $ref: "#/components/responses/HTTP400"
        "401":
          $ref: "#/components/responses/HTTP401"
        "403":
          $ref: "#/components/responses/HTTP403"
        "404":
          $ref: "#/components/responses/HTTP404"
      security:
        - OAuth2: []
        - BasicAuthentication: []
    delete:
      summary: Delete defined protected address by its ID
      description: >-
        Delete a particular protected address by its ID. The ID is defined as a
        path parameter and should be of ULID format.
      operationId: deletePrAddr
      tags:
        - Protected Addresses
      parameters: []
      responses:
        "204":
          description: ""
          content: {}
          headers: {}
        "401":
          $ref: "#/components/responses/HTTP401"
        "403":
          $ref: "#/components/responses/HTTP403"
        "404":
          $ref: "#/components/responses/HTTP404"
      security:
        - OAuth2: []
        - BasicAuthentication: []
    parameters:
      - in: path
        name: id
        description: "Protected Address ID"
        schema:
          type: string
        required: true
...
components:
  schemas:
...
    userData:
      type: object
      properties:
        login:
          type: string
          description: user login (for OIDC support should be formatted as email)
        first_name:
          type: string
          description: user first name
        last_name:
          type: string
          description: user last name
        type:
          type: string
          description: user type
        id:
          type: string
          description: user id
        failed_attempts:
          type: number
          description: number of failed authentication attempts
        lockout_until:
          type: string
          description: date/time the user account was lockout until
      required:
        - login
        - first_name
        - last_name
        - type
        - id
...
    protectedAddressData:
      type: object
      properties:
        email:
          type: string
          format: email
        owner:
          $ref: "#/components/schemas/userData"
        metadata:
          $ref: "#/components/schemas/addressMetadata"
        id:
          type: string
      required:
        - email
        - owner
        - metadata
        - id
...
    addressMetadata:
      type: object
      properties:
        comment:
          type: string
        service_name:
          type: string
      required:
        - comment
        - service_name
    error:
      type: object
      properties:
        id:
          type: number
          description: error id
        msg:
          type: string
          description: error description
      required:
        - id
        - msg
...
  securitySchemes:
    ApiToken:
      type: apiKey
      in: header
      name: X-OVOO-KEY
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://localhost.com/authorize
          tokenUrl: https://localhost.com/token
          scopes: {}
    BasicAuthentication:
      type: http
      scheme: basic
  requestBodies:
...
    createUserRequest:
      required: false
      description: ""
      content:
        application/json:
          schema:
            type: object
            properties:
              login:
                type: string
              first_name:
                type: string
              last_name:
                type: string
              type:
                type: string
              password:
                type: string
            required:
              - login
              - first_name
              - last_name
              - type
    updateUserRequest:
      required: false
      description: ""
      content:
        application/json:
          schema:
            type: object
            properties:
              first_name:
                type: string
              last_name:
                type: string
              type:
                type: string
    createProtectedAddressRequest:
      required: false
      description: ""
      content:
        application/json:
          schema:
            type: object
            properties:
              email:
                type: string
              metadata:
                $ref: "#/components/schemas/addressMetadata"
            required:
              - email
              - metadata
    updateProtectedAddressRequest:
      required: false
      description: ""
      content:
        application/json:
          schema:
            type: object
            properties:
              metadata:
                $ref: "#/components/schemas/addressMetadata"
...
  responses:
    HTTP403:
      description: Forbidden error response
      headers: {}
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error"
    HTTP404:
      description: Not found error response
      headers: {}
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error"
    HTTP401:
      headers: {}
      description: Unauthorized error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error"
    HTTP400:
      description: Bad or malformed request error response
      headers: {}
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error"
...
    getUsersResponse:
      headers: {}
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/userData"
      description: A list of users
    getUserDetailsResponse:
      description: Response containing detailed user data
      headers: {}
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/userData"
    createUserResponse:
      description: Response containing detailes on the created user
      headers: {}
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/userData"
    updateUserResponse:
      description: Response contains details on user after applied updates
      headers: {}
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/userData"
    getPrAddrsResponse:
      description: A list of protected addresses
      headers: {}
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/protectedAddressData"
    getPrAddrDetailsResponse:
      description: Response containing detailed protected address data
      headers: {}
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/protectedAddressData"
    createPrAddrResponse:
      description: Response containing detailes on the created protected address
      headers: {}
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/protectedAddressData"
    updatePrAddrResponse:
      description: Response contains details on protected address after applied updates
      headers: {}
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/protectedAddressData"
...